+++ CHANGELOG.rst
@@ -14,8 +14,8 @@ features.
 - Made :class:`brownie.datastructures.CombinedDict` hashable.
 - Made :class:`brownie.datastructures.ImmutableMultiDict` hashable.
 - Made :class:`brownie.datastructures.ImmutableOrderedDict` hashable.
-- Added :func:`brownie.functional.get_signature`.
 - Added :func:`brownie.Datastructures.namedtuple`.
+- Added :class:`brownie.functional.Signature`.
 
 0.4.1
 -----